# Faster Q2_K on Metal