Following in the footsteps of #2290 and #2294.

TG-128 in ms/t on M2 Max with 30-core GPU:

| Model | Master | This PR | Speedup |
|--:|--:|--:|--:|
| 7B | 22.5 | 18.4 | 22.3% |
|13B | 37.7 | 30.1 | 25.3% |
|33B | 88.9 | 68.7 | 29.4% |
|65B | 165.5 | 128.3 | 29.0% |